move AUtoComplete and AutoCompleteItem to beta module
