+++ stories/primer/beta/auto_complete/input_stories.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
-class Primer::AutoComplete::InputStories < ViewComponent::Storybook::Stories
+require "primer/beta/auto_complete"
+
+class Primer::Beta::AutoComplete::InputStories < ViewComponent::Storybook::Stories
   layout "storybook_preview"
 
   story(:auto_complete_input) do
